refactor: refactor the profile component
